requirements.txt
pytest
uiautomator2
telnetlib
telnetlib3
selenium
PyYAML
matplotlib
pandas
openpyxl
pytest-repeat
pytest-html
pytest-rerunfailures
psutil
pyserial
PyMySQL
cryptography
PyUserInput
pysnmp
roku
pytest-dependency